requirements.txt
python-telegram-bot==21.10
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1
alembic==1.14.1
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0
pandas==2.2.3
numpy
ta==0.11.0
matplotlib==3.9.3
mplfinance==0.12.10b0
plotly==5.24.1
kaleido==0.2.1
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
prophet
pytesseract
pillow
